+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.2.0] - 2026-03-30
+
 - ASB: Add limits to prevent denial-of-service via resource exhaustion.
 
 ## [4.1.1] - 2026-03-26
@@ -887,7 +889,8 @@ It is possible to migrate critical data from the old db to the sqlite but there 
 - Fixed an issue where Alice would not verify if Bob's Bitcoin lock transaction is semantically correct, i.e. pays the agreed upon amount to an output owned by both of them.
   Fixing this required a **breaking change** on the network layer and hence old versions are not compatible with this version.
 
-[unreleased]: https://github.com/eigenwallet/core/compare/4.1.1...HEAD
+[unreleased]: https://github.com/eigenwallet/core/compare/4.2.0...HEAD
+[4.2.0]: https://github.com/eigenwallet/core/compare/4.1.1...4.2.0
 [4.1.1]: https://github.com/eigenwallet/core/compare/4.1.0...4.1.1
 [4.1.0]: https://github.com/eigenwallet/core/compare/4.0.5...4.1.0
 [4.0.5]: https://github.com/eigenwallet/core/compare/4.0.4...4.0.5